Initial commit of notes on fixing Robin
Robin is an Android client for ADN that I use on a daily basis. It was
working perfectly well until I logged out from it (my 2016 Lenten
sacrifice to abstain from social media). Upon my return to ADN on Easter
Sunday imagine my surprise to find that the tokens were not working
(don't ask why, such things are beyond me). Luckily @33mhz mentioned
that the version @scruffyfox released worked fine...one thing led to
another...and here I am trying to fix the dashed thing so it works the
way I want it to. For now my scope is just to do this:

- fix current v1.apk so the text window when creating posts works
- fix current v2.apk so Dark Theme is selectable and works